# openapi.yaml
openapi: 3.0.3
info:
  title: Transparency Dashboard API
  description: API for tracking grants and spending on-chain for organizational transparency
  version: 1.0.0
  contact:
    name: Ethereum Community Foundation
servers:
  - url: /api/v1
    description: API server
paths:
...
  /api/v1/grants:
    get:
      summary: Get grants
      description: Returns a list of grants with optional filtering
      operationId: getGrants
      tags:
        - Grants
      parameters:
        - name: status
          in: query
          description: Filter by grant status (active/previous)
          required: false
          schema:
            type: string
            enum: [active, previous]
        - name: search
          in: query
          description: Search grants by name
          required: false
          schema:
            type: string
            example: "ethereum"
        - name: limit
          in: query
          description: Maximum number of grants to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of grants to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successfully retrieved grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grant'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create grant
      description: Creates a new grant
      operationId: createGrant
      tags:
        - Grants
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the grant
                  example: "Ethereum Research Grant"
                recipientName:
                  type: string
                  description: Name of the recipient
                  example: "Research Foundation"
                description:
                  type: string
                  description: Description of the grant
                  example: "Research into Ethereum scalability solutions"
                teamUrl:
                  type: string
                  format: uri
                  description: URL to the team's website
                  example: "https://example.com/team"
                  nullable: true
                projectUrl:
                  type: string
                  format: uri
                  description: URL to the project
                  example: "https://github.com/example/project"
                  nullable: true
                recipientAddress:
                  type: string
                  pattern: '^(0x)?[0-9a-f]{40}$'
                  description: Ethereum address of the recipient
                  example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
                totalGrantAmount:
                  type: string
                  description: Total grant amount (as string to preserve precision) in ETH
                  example: "100000.0"
                initialGrantAmount:
                  type: string
                  description: Initial grant amount (as string to preserve precision) in ETH
                  example: "25000.0"
                startDate:
                  type: string
                  format: date
                  description: Start date of the grant
                  example: "2023-12-01"
                expectedCompletionDate:
                  type: string
                  format: date
                  description: Expected completion date
                  example: "2024-06-01"
              required:
                - name
                - recipientName
                - description
                - recipientAddress
                - totalGrantAmount
                - initialGrantAmount
                - startDate
                - expectedCompletionDate
      responses:
        '201':
          description: Grant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/grants/{id}:
    put:
      summary: Update grant
      description: Updates grant details
      operationId: updateGrant
      tags:
        - Grants
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The grant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the grant
                  example: "Ethereum Research Grant"
                recipientName:
                  type: string
                  description: Name of the recipient
                  example: "Research Foundation"
                description:
                  type: string
                  description: Description of the grant
                  example: "Research into Ethereum scalability solutions"
                teamUrl:
                  type: string
                  format: uri
                  description: URL to the team's website
                  example: "https://example.com/team"
                  nullable: true
                projectUrl:
                  type: string
                  format: uri
                  description: URL to the project
                  example: "https://github.com/example/project"
                  nullable: true
                status:
                  type: string
                  description: Current status of the grant
                  example: "Milestone 2 completed"
                recipientAddress:
                  type: string
                  pattern: '^(0x)?[0-9a-f]{40}$'
                  description: Ethereum address of the recipient
                  example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
                totalGrantAmount:
                  type: string
                  description: Total grant amount (as string to preserve precision)
                  example: "100000.0"
                initialGrantAmount:
                  type: string
                  description: Initial grant amount (as string to preserve precision)
                  example: "25000.0"
                startDate:
                  type: string
                  format: date
                  description: Start date of the grant
                  example: "2023-12-01"
                expectedCompletionDate:
                  type: string
                  format: date
                  description: Expected completion date
                  example: "2024-06-01"
              required:
                - name
                - recipientName
                - description
                - recipientAddress
                - totalGrantAmount
                - initialGrantAmount
                - startDate
                - expectedCompletionDate
      responses:
        '200':
          description: Grant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/grants/{id}/milestones:
    put:
      summary: Update grant milestones
      description: Updates the milestones for a grant, including completion and sign-off status
      operationId: updateGrantMilestones
      tags:
        - Grants
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The grant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                milestones:
                  type: array
                  description: List of grant milestones
                  items:
                    $ref: '#/components/schemas/Milestone'
              required:
                - milestones
      responses:
        '200':
          description: Grant milestones updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/grants/{grantId}/disbursements:
    get:
      summary: Get grant disbursements
      description: Returns the disbursements for a specific grant
      operationId: getGrantDisbursements
      tags:
        - Grants
      parameters:
        - name: grantId
          in: path
          required: true
          description: The grant ID
          schema:
            type: string
            format: uuid
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '200':
          description: Successfully retrieved grant disbursements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Disbursement'
        '400':
          description: Bad request - invalid grant ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create disbursement
      description: Creates a new disbursement for a grant
      operationId: createDisbursement
      tags:
        - Grants
      security:
        - bearerAuth: []
      parameters:
        - name: grantId
          in: path
          required: true
          description: The grant ID
          schema:
            type: string
            format: uuid
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisbursementRequest'
      responses:
        '201':
          description: Disbursement created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Disbursement created successfully"
        '400':
          description: Bad request - invalid input or grant ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to fetch treasury data"
        code:
          type: string
          description: Error code
          example: "TREASURY_FETCH_ERROR"
      required:
        - error
        - code
...
    Grant:
      type: object
      properties:
        id:
          type: string
          description: Unique grant identifier
          example: "grant_12345"
        name:
          type: string
          description: Name of the grant
          example: "Ethereum Research Grant"
        recipientName:
          type: string
          description: Name of the recipient
          example: "Research Foundation"
        description:
          type: string
          description: Description of the grant
          example: "Research into Ethereum scalability solutions"
        teamUrl:
          type: string
          format: uri
          description: URL to the team's website
          example: "https://example.com/team"
          nullable: true
        projectUrl:
          type: string
          format: uri
          description: URL to the project
          example: "https://github.com/example/project"
          nullable: true
        status:
          type: string
          description: Current status of the grant
          example: "Milestone 2 completed"
        recipientAddress:
          type: string
          pattern: '^(0x)?[0-9a-f]{40}$'
          description: Ethereum address of the recipient
          example: "0x742d35cc6670c02bb4bbecbb5fb95e8b7d7b36bf"
        totalGrantAmount:
          type: string
          description: Total grant amount (as string to preserve precision)
          example: "100000.0"
        initialGrantAmount:
          type: string
          description: Initial grant amount (as string to preserve precision)
          example: "25000.0"
        startDate:
          type: string
          format: date
          description: Start date of the grant
          example: "2023-12-01"
        expectedCompletionDate:
          type: string
          format: date
          description: Expected completion date
          example: "2024-06-01"
        disbursements:
          type: array
          description: List of grant disbursements
          items:
            $ref: '#/components/schemas/Disbursement'
        fundsUsage:
          type: array
          description: List of funds usage entries showing where grant funds have been used
          items:
            $ref: '#/components/schemas/FundsUsage'
        amountGivenSoFar:
          type: string
          description: Total amount disbursed so far
          example: "50000.0"
        milestones:
          type: array
          description: List of grant milestones
          items:
            $ref: '#/components/schemas/Milestone'
      required:
        - id
        - name
        - recipientName
        - description
        - recipientAddress
        - totalGrantAmount
        - initialGrantAmount
        - startDate
        - expectedCompletionDate
        - disbursements
        - fundsUsage
        - amountGivenSoFar
        - milestones
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...